Fix all relevant GCC warnings (OOS Open Beta 6)
This is the second part in the GCC warnings series, fixing all
warnings that come from OnePlus modifications to CAF.

This was done with the following toolchain version:
arch64-linux-gnu-gcc (Linaro GCC 7.2-2017.11) 7.2.1 20171011

If a warning was a false positive, it was disabled in the folder's
Makefile. I believe all of these are correct but I am 100% open to
discussion (I am not infallible).

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>